+++ src/main/java/repository/dsm/OriginDsmRepository.java
@@ -1,19 +1,17 @@
 package repository.dsm;
 
 import config.JdbcTemplate;
-
+import domain.SqlReader;
 import java.io.File;
 import java.sql.Connection;
 import java.sql.SQLException;
-import domain.SqlReader;
 import repository.ExecuteQuery;
-import repository.dsm.SaveDsmTemp;
 
-public class DsmRepository {
+public class OriginDsmRepository {
     private JdbcTemplate jdbcTemplate = new JdbcTemplate();
     private ExecuteQuery executeQuery = new ExecuteQuery();
 
-    public void saveOriginData(SqlReader createSql, File[] dsms) {
+    public void run(SqlReader createSql, File[] dsms) {
         try (Connection conn = jdbcTemplate.getConnection()) {
             createTable(conn, createSql);
             saveDsmTemp(conn, dsms);
@@ -38,29 +36,4 @@ public class DsmRepository {
         executeQuery.delete(conn, sql);
     }
 
-    public void procOriginData(SqlReader createSql) {
-        try (Connection conn = jdbcTemplate.getConnection()) {
-            createTable(conn, createSql);
-            dropDsmTempTable(conn);
-            setPrimaryKey(conn);
-            createIndex(conn);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void dropDsmTempTable(Connection conn) {
-        String sql = "DROP table dsm_temp";
-        executeQuery.drop(conn, sql);
-    }
-
-    private void setPrimaryKey(Connection conn) {
-        String sql = "ALTER TABLE dsm ADD COLUMN id SERIAL PRIMARY KEY;";
-        executeQuery.alter(conn, sql);
-    }
-
-    private void createIndex(Connection conn) {
-        String sql = "CREATE INDEX dsm_sig_cd_index ON dsm USING brin(sig_cd);";
-        executeQuery.createIndex(conn, sql);
-    }
 }
